+++ app/src/main/java/com/example/movieaddiction/MainActivity.kt
@@ -1,7 +1,9 @@
 package com.example.movieaddiction
 
+import androidx.activity.compose.setContent
+import androidx.navigation.compose.rememberNavController
 import com.example.movieaddiction.databinding.ActivityMainBinding
-import com.example.movieaddiction.ui.WelcomeFragment
+import com.example.movieaddiction.ui.composables.navigation.AppNavigation
 import com.google.accompanist.pager.ExperimentalPagerApi
 import com.karleinstein.basemvvm.base.BaseActivity
 import com.karleinstein.basemvvm.extension.addFragment
@@ -12,10 +14,10 @@ class MainActivity : BaseActivity() {
     override val viewBinding: ActivityMainBinding by viewBinding(ActivityMainBinding::inflate)
 
     override fun bindView() {
-        supportFragmentManager.addFragment(
-            fragment = WelcomeFragment.newInstance(),
-            container = R.id.root_content_view
-        )
+        setContent {
+            val navHostController = rememberNavController()
+            AppNavigation(navController = navHostController)
+        }
     }
 
     override fun setUpView() {
