+++ app/src/main/java/com/example/movieaddiction/di/AppModule.kt
@@ -1,7 +1,6 @@
 package com.example.movieaddiction.di
 
-import androidx.lifecycle.viewmodel.compose.viewModel
-import com.example.movieaddiction.ui.WelcomeViewModel
+import com.example.movieaddiction.ui.composables.viewmodels.WelcomeViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
