+++ app/src/main/java/com/example/movieaddiction/ui/composables/navigation/MovieAddictionScreen.kt
@@ -1,0 +1,7 @@
+package com.example.movieaddiction.ui.composables.navigation
+
+enum class MovieAddictionScreen {
+    LOGIN_ROUTE,
+    WELCOME_ROUTE,
+    HOME_ROUTE
+}
