+++ app/src/main/java/com/example/movieaddiction/ui/composables/screen/HomeScreen.kt
@@ -1,0 +1,11 @@
+package com.example.movieaddiction.ui.composables.screen
+
+import androidx.compose.runtime.Composable
+import androidx.navigation.NavHostController
+
+@Composable
+fun HomeScreen(
+    navHostController: NavHostController
+) {
+
+}
