+++ app/src/main/java/com/example/movieaddiction/ui/composables/viewmodels/WelcomeViewModel.kt
@@ -1,11 +1,12 @@
-package com.example.movieaddiction.ui
+package com.example.movieaddiction.ui.composables.viewmodels
 
+import com.example.movieaddiction.utils.Karl
 import com.karleinstein.basemvvm.base.BaseViewModel
 
 class WelcomeViewModel : BaseViewModel() {
 
     fun onPagerChange(page: Int) {
-
+        Karl.d("page change: $page")
     }
 
     fun calculateDirection(startPoint: Float, endPoint: Float) {
